move fixtures into src package in order to have type safety
